requirements.txt
praw>=7.7.0
fpdf2>=2.7.0
pyyaml>=6.0
python-dotenv>=1.0.0
